test(vitest): migrate browser API tests and remove jsdom polyfills
