+++ packages/react/src/core/system/use-system-color-mode.test.browser.tsx
@@ -1,4 +1,4 @@
-import { renderHook } from "#test"
+import { renderHook } from "#test/browser"
 import { useSystemColorMode } from "./use-system-color-mode"
 
 const mockMatchMedia = (matches = false) => {
@@ -22,34 +22,34 @@ describe("useSystemColorMode", () => {
     mockMatchMedia(false)
   })
 
-  test("returns system color mode", () => {
-    const { result } = renderHook(() =>
+  test("returns system color mode", async () => {
+    const { result } = await renderHook(() =>
       useSystemColorMode({ defaultColorMode: "light" }),
     )
     expect(result.current.systemColorMode).toBeDefined()
     expect(["light", "dark"]).toContain(result.current.systemColorMode)
   })
 
-  test("getSystemColorMode returns light when prefers light", () => {
+  test("getSystemColorMode returns light when prefers light", async () => {
     mockMatchMedia(false)
-    const { result } = renderHook(() =>
+    const { result } = await renderHook(() =>
       useSystemColorMode({ defaultColorMode: "light" }),
     )
     expect(result.current.getSystemColorMode()).toBe("light")
   })
 
-  test("getSystemColorMode returns dark when prefers dark", () => {
+  test("getSystemColorMode returns dark when prefers dark", async () => {
     mockMatchMedia(true)
-    const { result } = renderHook(() =>
+    const { result } = await renderHook(() =>
       useSystemColorMode({ defaultColorMode: "light" }),
     )
     expect(result.current.getSystemColorMode()).toBe("dark")
   })
 
-  test("accepts callback option without error", () => {
+  test("accepts callback option without error", async () => {
     const callback = vi.fn()
 
-    renderHook(() =>
+    await renderHook(() =>
       useSystemColorMode({ callback, defaultColorMode: "light" }),
     )
   })
